Create a UBossAbility class that is a copy of the UAbility class
I want to have some sort of Ability class in the plugin that serves as a
base class for all of the boss' abilities. The UAbility class seems to
be a great candidate, but I really don't want to have the plugin depend
on the host project (as plugins shouldn't do that). So we are going to
have some duplicated classes in here, but it is for the best.